+++ packages/core/src/lib/state.ts
